% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{getMetricsTable}
\alias{getMetricsTable}
\alias{getMetricsTable,ObjXAI-method}
\alias{setMetricsTable,ObjXAI-method}
\title{Get the Metrics Table}
\usage{
getMetricsTable(object)

\S4method{getMetricsTable}{ObjXAI}(object)
}
\arguments{
\item{object}{An ObjXAI object.}
}
\value{
A data frame containing the metrics.
}
\description{
This method retrieves the metrics table from an ObjXAI object.
}
