% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ObjXAI-class}
\alias{ObjXAI-class}
\alias{.objXAI}
\title{ObjXAI class}
\value{
A ObjXAI object
}
\description{
\code{ObjXAI} is a class used to store the output values
of the \code{XAI.test} function.
}
\examples{

obj <- new("ObjXAI", 
           data = data.frame(), 
           dataSim = data.frame(), 
           metricsTable = data.frame(Metric = c("Accuracy", "Precision"), 
                                       Value = c(0.95, 0.89)), 
           map = list(), 
           models = list(), 
           modelPredictions = list(), 
           args = list())

}
