% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{show,ObjXAI-method}
\alias{show,ObjXAI-method}
\title{Show Method for ObjXAI}
\usage{
\S4method{show}{ObjXAI}(object)
}
\arguments{
\item{object}{An ObjXAI object.}
}
\description{
Prints the first 5 rows of the metrics table from an ObjXAI object.
}
