% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{getMetricsTable}
\alias{getMetricsTable}
\title{Get the Metrics Table}
\usage{
getMetricsTable(object)
}
\arguments{
\item{object}{An ObjXAI object.}
}
\value{
A data frame containing the metrics.
}
\description{
This method retrieves the metrics table from an ObjXAI object.
}
\examples{

obj <- new("ObjXAI", 
           data = data.frame(), 
           dataSim = data.frame(), 
           metricsTable = data.frame(Metric = c("Accuracy", "Precision"), 
                                       Value = c(0.95, 0.89)), 
           map = list(), 
           models = list(), 
           modelPredictions = list(), 
           args = list())
getMetricsTable(obj)

}
