% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{objXAI-class}
\alias{objXAI-class}
\alias{contXAI}
\title{objXAI class}
\value{
A objXAI object
}
\description{
\code{objXAI} is a class used to store the output values
of the \code{XAI.test} function.
}
\examples{

a=1
}
