% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{setMetricsTable}
\alias{setMetricsTable}
\alias{setMetricsTable,ObjXAI-method}
\title{Set the Metrics Table}
\usage{
setMetricsTable(object, value)

\S4method{setMetricsTable}{ObjXAI}(object, value)
}
\arguments{
\item{object}{An ObjXAI object.}

\item{value}{A data frame to set as the metrics table.}
}
\value{
The modified ObjXAI object.
}
\description{
This method sets the metrics table for an ObjXAI object.
}
